+++ bot.py
@@ -9,104 +9,20 @@ config = dotenv_values(".env")
 client = commands.Bot(command_prefix=".")
 
 
+from stopwatch import stopwach
+from pomodoro import pomodoro
+from daily_remainder import daily_reminder
+from hourly_reminder import hourly_reminder
+from alarm import alarm
+from timer import start_timer
+
+
 @client.event
 async def on_ready():
     print("Bot is ready")
 
 
-@client.command(aliases=["timer", "start-timer"])
-async def start_timer(ctx, seconds):
 
-    if not seconds.isnumeric():
-        await ctx.send("Must be a number!")
-
-    # Convert string (user input command) into an integer
-    seconds = int(seconds)
-    if seconds <= 0:
-        await ctx.send("I don't think I'm allowed to do negatives")
-
-    # Send a message to the Discord channel
-    await ctx.send(f"Your countdown has been started for {str(seconds)} seconds")
-
-    # We are making a variable, so that we can reference our message later on
-    # and thus, edit our message on the Discord Server
-    message = await ctx.send(f"Time Remaining : {str(seconds)} second(s)")
-
-    while True:
-        seconds -= 1
-        if seconds == 0:
-            await message.edit(content=(f"Time Remaining : 0"))
-
-            # Notify the author (user who issued the command) using ctx.message.author.mention
-            await ctx.send(ctx.message.author.mention + " Your countdown has ended!")
-            break
-
-        # Thus, we edit our message on the Discord channel with the current remaining time
-        await message.edit(content=(f"Time Remaining : {str(seconds)} second(s)"))
-
-        # Sleep for 1 second
-        await asyncio.sleep(1)
-
-
-@client.command(aliases=["set-alarm"])
-async def alarm(ctx, time):
-    # Take user input (time in 24-hour format), for example 23:00
-    # Check if it is a valid 24-hour format, and return an apppropriate message
-    # Parse the user input using time.split(":")
-    # Remind the user as soon as it's time
-    pass
-
-
-@client.command(aliases=["stopwatch", "start-stopwatch"])
-async def stopwach(ctx):
-     
-    
-    pass
-
-
-@client.command(aliases=["hourly-reminder", "set-hourly-reminder"])
-async def hourly_reminder(ctx, task):
-    # Takes input from the user (task) about what they would like to accomplish
-    # Send a reminder after one hour to see how far they have come up with the task
-    # At the end of every hour, ask the user if they have completed the task.
-    # Wait for the stop command from user (You may use "client.wait_for")
-    pass
-
-
-@client.command(aliases=["daily-reminder", "set-daily-alarm"])
-async def daily_reminder(ctx, time):
-    # Takes input from the user (task) about what they would like to accomplish
-    # Send a reminder after 24 hours to see how far they have come up with the task
-    # At the end of every hour, ask the user if they have completed the task.
-    # Wait for the stop command from user (You may use "client.wait_for")
-    pass
-
-
-@client.command(aliases=["start-pomodoro"])
-async def pomodoro(ctx, time):
-    print("Pomodoro starts now!")
-    for i in range(4):
-        t=25*60
-        while t:
-            mins = t // 60
-            secs = t%60
-            timer = '{:02d}:{:02d}'.format(mins,secs)
-            print(timer, end="\r")
-            time.sleep(1)
-            t -=1
-            print("Break time!!")
-            t= 5*60
-            while t:
-                mins = t//60
-                secs = t%60
-                timer= '{:02d}:{:02d}'.format(mins,secs)
-                print(timer, end="\r")
-                time.sleep(1)
-                t -=1
-                print("Work time!!")
-
-   
-    pass
 
 
 def getQuote(tags=["education", "success"]):  # default arguments
